+++ examples/Macro/LogBench.php
@@ -10,17 +10,18 @@
  *
  */
 
-namespace PhpBench\Benchmarks\Macro;
+namespace PhpBench\Examples\Macro;
 
 use PhpBench\Console\Command\LogCommand;
 use PhpBench\Console\Command\RunCommand;
+use PhpBench\Examples\Macro\BaseBenchCase;
 
 
 
 /**
  * Benchmark for the log command.
  *
- * @BeforeClassMethods({"createWorkspace", "createResults"})
+ * @BeforeClassMethods({"resetWorkspace", "createResults"})
  */
 class LogBench extends BaseBenchCase
 {