+++ examples/Macro/RunBench.php
@@ -10,9 +10,10 @@
  *
  */
 
-namespace PhpBench\Benchmarks\Macro;
+namespace PhpBench\Examples\Macro;
 
 use PhpBench\Console\Command\RunCommand;
+use PhpBench\Examples\Macro\BaseBenchCase;
 
 
 /**